Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/workers/dedicated-worker-parse-error-failure-expected.txt

CONSOLE MESSAGE: SyntaxError: Unexpected token ';'
CONSOLE MESSAGE: SyntaxError: Unexpected token '*'. import call expects exactly one argument.

FAIL Classic worker construction for script with syntax error should dispatch an event named error. assert_equals: expected function "function Event() {
    [native code]
}" but got function "function ErrorEvent() {
    [native code]
}"
FAIL Static import on classic worker should dispatch an event named error. assert_equals: expected function "function Event() {
    [native code]
}" but got function "function ErrorEvent() {
    [native code]
}"
